+++ src/RestApi/MapRestController.php
@@ -58,6 +58,7 @@ class MapRestController extends AbstractFOSRestController
             $excludedStoreTypes = [];
             $teamSearchStatus = [];
             $status = $paramFetcher->get('status');
+            $userId = null;
             if (is_array($status) && !empty($status)) {
                 foreach ($status as $s) {
                     switch ($s) {
@@ -72,11 +73,14 @@ class MapRestController extends AbstractFOSRestController
                                 CooperationStatus::COOPERATION_STARTING, CooperationStatus::COOPERATION_ESTABLISHED
                             ]);
                             break;
+                        case 'mine':
+                            $userId = $this->session->id();
+                            break;
                     }
                 }
             }
 
-            $markers['betriebe'] = $this->storeGateway->getStoreMarkers($excludedStoreTypes, $teamSearchStatus);
+            $markers['betriebe'] = $this->storeGateway->getStoreMarkers($excludedStoreTypes, $teamSearchStatus, $userId);
         }
 
         return $this->handleView($this->view($markers, 200));