+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - New UI for Stores information and modification of stores !2645 !2685 !2688 @KoeMai
 - Added turkish translation !2720 @Evren
 - Store chains can now be manged from the website. The working group "Betriebsketten" contains a tabular overview of all chains. !2251 !2273 @AntonBallmaier
+- Store markers on the map can be filtered by the user's membership #1025 !2682 @alex.simm
 
 ## Changes
 - Make menubar sticky on mobile !2411 @AntonBallmaier